Addresses issue #338. Adds the ability to create a Map and OrderedMap from a List of Lists (if this is something that should actually be supported). Ran the perf tests and all seems to be good.