@leebyron 2015-03-12
There's no perf test covering this case yet. We will first need good perf coverage of this use case. I imagine creating a new Seq on every iteration is probably quite expensive.