@ellbee 2015-03-13
Re the perf tests, I was just trying to make sure that I hadn't caused a performance regression for existing cases as I had the same worries about creating a Seq each time vs straight up array access. I'll get perf coverage added, I guess it should follow the 2, 8, 32, 1024 pattern?